requirements.txt
Django==3.2.8
Pillow==8.4.0
tensorflow==2.6.0